Update README.md
JavaScript Projects

Here are some projects I have created using Javascript

Pizza_Project
Calculator
TicTacToe

Pizza Project
This project is a menu used to order pizza with a various selection of sizes and toppings.
It takes the information and calculates the price for the order.

Tic-Tac-Toe
A clasic game between you and the computer.

Calculator 
A basic 4 function calculator. You can add, subtract, divide, and multiply.